pyproject.toml
[tool.poetry]
name = "django-saas-template"
version = "0.1.0"
description = ""
authors = ["Jamal Moir <jamal.moir@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
dj-stripe = "^2.6.1"
sentry-sdk = "^1.5.8"
decouple = "^0.0.7"
stripe = "^2.69.0"
djangorestframework-simplejwt = "^5.1.0"
django-cors-headers = "^3.11.0"
django-extensions = "^3.1.5"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
ipdb = "^0.13.9"
pytest = "^7.1.1"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
